requirements.txt
anthropic>=0.49.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
